projects/21-saacgnet/requirements.txt
Flask
flask-cors
flask-sqlalchemy
SQLAlchemy
pandas
openpyxl
psycopg2-binary
gunicorn
python-dotenv
